% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_preview}
\alias{df_preview}
\title{df_preview}
\usage{
df_preview(a_df, n_rows = 300)
}
\arguments{
\item{a_df}{a data.frame or compatible object.}

\item{n_rows}{a number of rows to sample for the preview.}
}
\value{
a data.frame with n_rows
}
\description{
Preview a data frame (or tibble)
}
\details{
Tibble has some issues with its pillar output and Emacs buffers. This
function will sample a data frame or tibble and return a plain data.frame.
}
