% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drive_download_dir}
\alias{drive_download_dir}
\title{Download a folder from Google Drive}
\usage{
drive_download_dir(drive_folder = NULL, dl_path = getwd())
}
\arguments{
\item{drive_folder}{a google drive folder path or id}

\item{dl_path}{a local path to dowload the folder to. Defaults to getwd().}
}
\value{
nothing.
}
\description{
Download a folder from Google Drive
}
\examples{
\dontrun{
drive_download_dir(
drive_folder = as_id("https://drive.google.com/open?id=0B7688WPR38x2Nk5yYV9ZMVYwWVE"),
dl_path = "./models")
}
}
