% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drive_fetch_all}
\alias{drive_fetch_all}
\title{Source all drive data fetch scripts in current directory.}
\usage{
drive_fetch_all()
}
\value{
nothing.
}
\description{
Source all drive data fetch scripts in current directory.
}
