% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_drive_fetch}
\alias{generate_drive_fetch}
\title{Generate an R file to fetch a folder or file from Google drive}
\usage{
generate_drive_fetch(a_link)
}
\arguments{
\item{a_link}{the google drive link}
}
\value{
nothing.
}
\description{
This function takes a link to a Google drive file or folder. If one is not supplied it will
attempt to read one from the clipboard. Given a Google drive link, an R script to fetch the content
is generated. The script is slightly different depending on whether the link points to a single file
or folder. The R script is generated in the current working directory '.' and will download files to the
current working directory '.'.
}
