% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uitls.R
\name{how_to}
\alias{how_to}
\title{How to use a package}
\usage{
how_to(package)
}
\arguments{
\item{package}{a package to open vigettes of.}
}
\value{
Vignette data, invisibly.
}
\description{
Open all package vignettes in browser tabs ready for skimming.
}
\examples{
\dontrun{
how_to("naniar")
}
}
