% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{howto}
\alias{howto}
\title{How to use a package}
\usage{
howto(package)
}
\arguments{
\item{package}{a package to open vigettes of.}
}
\value{
Vignette data, invisibly.
}
\description{
Open all package vignettes in browser tabs ready for skimming.
}
\examples{
\dontrun{
howto("naniar")
}
}
