% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_clipboard}
\alias{install_clipboard}
\title{Install the clipboard}
\usage{
install_clipboard()
}
\value{
nothing
}
\description{
install_clipboard
}
\details{
Assuming the clipboard is a series of library(pkg) calls, install all of them.
}
