% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_ghzip}
\alias{install_ghzip}
\title{install_ghzip}
\usage{
install_ghzip(name_repo)
}
\arguments{
\item{name_repo}{a github repo identifier, name/repo e.g. "milesmcbain/mmmisc"}
}
\value{
nothing. Installs package.
}
\description{
install a githup repo using the master.zip
}
\details{
This is useful if you find yourself trapped behind a man in the middle proxy that is stuffing up your ssl traffic.
}
