% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_missing_deps}
\alias{install_missing_deps}
\title{install_missing_deps}
\usage{
install_missing_deps(dep_file)
}
\value{
nothing. Installs as side effect.
}
\description{
Detect dependencies in a file and install those that are missing
}
\author{
Miles McBain
}
