% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_R_file}
\alias{is_R_file}
\title{Test if a file has an rmd or R extension.}
\usage{
is_R_file(filename)
}
\arguments{
\item{filename}{of a file. Can include path.}
}
\value{
A logical value. True if the file ends in
.r, .rmd or any mixed-case variant of those.
}
\description{
Test if a file has an rmd or R extension.
}
