% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate_renv}
\alias{isolate_renv}
\title{isolate_renv()}
\usage{
isolate_renv()
}
\value{
nothing
}
\description{
Set your libPaths to .libPaths()[[1]]
}
\details{
If you're using 'renv' this will hopefully remove your system libraries and give you an isolated environment.
}
