% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_ifelse}
\alias{map_ifelse}
\title{Map if predicate, otherwise take a value.}
\usage{
map_ifelse(.x, .p, .f, .e, ...)
}
\arguments{
\item{.x}{a list}

\item{.p}{a list of logicals or predicate function to evaluate on list}

\item{.f}{a function}

\item{.e}{a value to use in results for false predicates.}

\item{...}{}
}
\value{
This applied to all elements: if(.p) .f(.x) else .e
}
\description{
Like map_if but with the option to specify a value for false cases.
}
