\name{MMmisc-package}
\alias{MMmisc-package}
\alias{MMmisc}
\docType{package}
\title{
Various functions which Mike wants to have around.
}
\description{
Various functions which Mike wants to have around.
}
\details{
Please look at the package index (linked at the foot of this page) for details of the contents.
}

\author{
Mike Meredith
}

