% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_as_per}
\alias{order_as_per}
\title{Order levels in factor as per some example string}
\usage{
order_as_per(fctr, order_chars)
}
\arguments{
\item{fctr}{a factor with levels}

\item{order_chars}{a character vector with strings matching levels of fctr.}
}
\value{
a version of `fctr` with levels ordered as per `order_chars`
}
\description{
This is the way fct_relevel should work. The warning is silly.
}
