% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{P0PE}
\alias{P0PE}
\title{Paste0, Parse, Eval}
\usage{
P0PE(..., .env = parent.frame())
}
\arguments{
\item{...}{text to parse and evaluate}

\item{env}{an environment to evaulate parsed expressions in. Defaults to parent.frame().}
}
\value{
the result of the evaluation.
}
\description{
Wraps up the common task for evaluting machine generated code.
Be careful with pipes (%>%) as these mess with the
evironment chain for eval(). Choose a suitable argument for .env
like .GlobalEnv or a binding to sys.frame().
}
