% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_match_except}
\alias{regex_match_except}
\title{Return characters not matching a regular expression}
\usage{
regex_match_except(text, pattern)
}
\arguments{
\item{text}{a vector of characer strings to be (un)matched.}

\item{pattern}{a regular expression.}
}
\value{
vector of substrings of `text` that dit not match the `pattern`.
}
\description{
Given a vector of chracter strings, this function returns portions of the stirng
that do not march a pattern, as a vector. This allows a chainable regex
filtering style that is readily compatible with `%>%`.
All matches from the resulting regexec call are pasted together. This will create
weird behaviour if you use capturing groups. Best to use non-capturing `(?:)`.
}
