% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_sub}
\alias{regex_sub}
\title{Substitute characters matching a regular expression.}
\usage{
regex_sub(text, pattern, replacement, sub_all = TRUE)
}
\arguments{
\item{text}{a vector of character strings to have patterns substituted.}

\item{pattern}{a string pattern that, if matched in `text` will be substituded with `replacement`.}

\item{replacement}{either a string or vector of strings to replace matching patterns in `text`.}

\item{sub_all}{Determined if all matches are substited (default), or the first ocurrence per input string.}
}
\value{
a vector of character strings
}
\description{
Given a vector of character strings, substitute parts of the strings
matching `pattern` with `replacement`. `replacement` can be a vector
argument, in which case it must be the same length as `text`, and
the substiuted `text` is made per `replacement` string.
}
\details{
The default mode is to substitute all pattern matches. This can be
switched to first match only by setting `sub_all = FALSE`.
}
\examples{
\dontrun{
regex_sub(c("ABC", "AXYZ", "DEFA"),"A", "_")
regex_sub(c("ABC_", "AXY_", "DEF_"),"_", c(1,2,3)
}

}
