% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rend}
\alias{rend}
\title{Render an Rmd file in directory}
\usage{
rend(path = ".", format = "all")
}
\arguments{
\item{path}{The path to search for .Rmd files.}

\item{format}{The output formats. The default is `"all"`, all formats listed in the file.}
}
\value{
Nothing.
}
\description{
This function is a replacement for the 'knit' from RStudio, for cases where you're not
working in RStudio.
}
\details{
Searches for all Rmd files below the `path` and prompts for
a selection of one to be rendered. File is rendered using default render formats,
or as specified in `format`. Once the file is rendered, it is opened in the browser.
Opening in browser will fail for .docx.
}
