% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_like}
\alias{rm_like}
\title{Remove variables with names like this}
\usage{
rm_like(pattern = character())
}
\arguments{
\item{pattern}{a regex pattern to match. Could be a simple
prefix or suffix string.}
}
\value{
Nothing.
}
\description{
Remove things in the global environment with name like a
string you supply. Like is in the regex match sense, so
if you make a slip blowing away most of your environment
is a possibility.
}
\examples{
\dontrun{
lm_1 = 1
lm_2 = 2
glm_1 = 3
rm_like("^lm_")

> ls()
[1] "glm_1"
}
}
