% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rs_get_symbol_at_cursor}
\alias{rs_get_symbol_at_cursor}
\title{rs_get_symbol_at_cursor}
\usage{
rs_get_symbol_at_cursor(context)
}
\arguments{
\item{context}{probably rstudioapi::getActiveDocumentContext()}
}
\value{
the symbol the cursor is on as text
}
\description{
Get the symbol at the rstudio cursor
}
\details{
to do something with it. I originally implemented this in drake.
}
\author{
Miles McBain
}
