% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{save_plot}
\usage{
save_plot(filename, a_plot)
}
\arguments{
\item{filename}{to save}

\item{a_plot}{a plot object}
}
\value{
nothing
}
\description{
quick plot save
}
\details{
save a plot with ragg
}
\author{
Miles McBain
}
