% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_lib_paths}
\alias{set_lib_paths}
\title{set_lib_paths}
\usage{
set_lib_paths(lib_vec)
}
\value{
nothing
}
\description{
Set your R library paths at run time
}
\details{
A wrapper for .libPaths that removes .Library and .Library.site from your
library paths while adding the contents of the character vector `lib_vec`.
}
