% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{webdocs}
\alias{webdocs}
\title{Open URL for a package from it's description file.}
\usage{
webdocs(package_name)
}
\arguments{
\item{package_name}{the package to open URL for}
}
\value{
nothing
}
\description{
The URL is fetched from CRAN, so the package need not be installed to run this function.
}
\examples{
\dontrun{
webdocs(visdat)
}

}
