% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whelp}
\alias{whelp}
\title{Web Help}
\usage{
whelp(search_term)
}
\arguments{
\item{search_term}{to be searched for on rdocumentation.org}
}
\value{
A browser tab with search results
}
\description{
Web Help
}
\examples{
\dontrun{
whelp(tribble_paste)
}
}
