\name{MMmisc-package}
\alias{MMmisc-package}
\alias{MMmisc}
\docType{package}
\title{
Various functions which Mike wants to have around.
}
\description{
Various functions which Mike wants to have around.
}
\details{
\tabular{ll}{
Package: \tab MMmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-15\cr
License: \tab GPL-3\cr
}
}
\author{
Mike Meredith
}

